// src/Models/OpcionesCONAFE.ts
// Generated by https://quicktype.io

export interface OpcionesCONAFE {
	ERROR_API:             string;
	MENSAJE_API:           string;
	VALID_TOKEN:           string;
	DISPOSITIVO_ACADEMICO: DispositivoAcademico[];
	OPCION_EQUIPO:         DispositivoAcademico[];
	OPCION_PAGO:           DispositivoAcademico[];
	TIPO_ESTUDIO:          DispositivoAcademico[];
	TIPO_FIGURA:           DispositivoAcademico[];
}

export interface DispositivoAcademico {
	VALUE: string;
	TEXT:  string;
}